Chap4 RaysCameraBackGround

+++ RayTracingGPU/vec3.h
@@ -131,12 +131,12 @@ __host__ __device__ inline vec3 operator*(const vec3& v1, const vec3& v2)
 	return vec3(v1.e[0] * v2.e[0], v1.e[1] * v2.e[1], v1.e[2] * v2.e[2]);
 }
 
-__host__ __device__ inline vec3 operator*(const vec3& v1, int t)
+__host__ __device__ inline vec3 operator*(const vec3& v1, double t)
 {
 	return vec3(v1.e[0] * t, v1.e[1] * t, v1.e[2] * t);
 }
 
-__host__ __device__ inline vec3 operator*(int t, const vec3& v1)
+__host__ __device__ inline vec3 operator*(double t, const vec3& v1)
 {
 	return v1 * t;
 }
@@ -146,7 +146,7 @@ __host__ __device__ inline vec3 operator/(const vec3& v1, const vec3& v2)
 	return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-__host__ __device__ inline vec3 operator/(const vec3& v1, int t)
+__host__ __device__ inline vec3 operator/(const vec3& v1, double t)
 {
 	return vec3(v1.e[0] / t, v1.e[1] / t, v1.e[2] / t);
 }
